+++ simplified_quad_dynamics.hpp
@@ -5,4 +5,4 @@
 
 using namespace Eigen;
 
-StateSpace quadruped_state_space_continuous(const float&  yaw, Matrix<float, 4, 3>& foot_positions);
+StateSpace quadruped_state_space_continuous(const double&  yaw, Matrix<double, 4, 3>& foot_positions);
